requirements.txt
tornado
selenium
markdown
watchdog
pillow
openai-functions

pypdf
nltk
spellchecker